19: Moving Container and Section component to globals
