+++ src/components/globals/index.tsx
@@ -1,3 +1,4 @@
+import { breakpoints } from '@constants/breakpoints';
 import styled from 'styled-components';
 
 export const Div = styled.div``;
@@ -10,7 +11,6 @@ export const H4 = styled.h4``;
 export const H5 = styled.h5``;
 export const H6 = styled.h6``;
 export const P = styled.p``;
-
 export const Strong = styled.strong`
   font-weight: bold;
 `;
@@ -26,4 +26,21 @@ export const BulletedList = styled.ul`
       margin-bottom: 0;
     }
   }
+`;
+
+export const Container = styled.div`
+  margin-right: auto;
+  margin-left: auto;
+  padding-right: 0.5rem;
+  padding-left: 0.5rem;
+  max-width: 1008px;
+
+  @media ${breakpoints.desktop} {
+    padding: 0;
+  }
+`;
+
+export const Section = styled.div`
+  padding-top: 3rem;
+  padding-bottom: 3rem;
 `;
