--- src/components/primitives/Section.tsx
@@ -1,6 +1,0 @@
-import styled from 'styled-components';
-
-export const Section = styled.div`
-  padding-top: 3rem;
-  padding-bottom: 3rem;
-`;
